working init and seed from top level directory
