+++ package.json
@@ -13,8 +13,7 @@
   "scripts": {
     "ufa:dev": "turbo run ufa:dev --ui=tui",
     "ufa:dev:clean": "turbo run ufa:dev:clean",
-    "ufa:dev:seed": "./ufa/scripts/dev-seed.sh",
-    "ufa:dev:sqlserver": "turbo run ufa:dev --filter=analytical-base --filter=vite-web-base --filter=transactional-sqlserver --filter=retrieval-base --ui=tui"
+    "ufa:dev:seed": "turbo run ufa:dev:seed"
   },
   "devDependencies": {
     "@turbo/gen": "^1.13.4",
