+++ turbo.json
@@ -12,7 +12,13 @@
     "lint": {},
     "ufa:dev": {
       "cache": false,
-      "persistent": true
+      "persistent": true,
+      "dependsOn": ["ufa:dev:setup"]
+    },
+
+    "ufa:dev:setup": {
+      "cache": false,
+      "persistent": false
     },
     "ufa:dev:clean": {
       "cache": false,
